Klasse Allgemein hinzugefügt
